Diálogo elimación de sesión. Barra superior con opciones. Navegación pantallas secundarias integrada.

+++ app/src/main/java/com/gym/ui/features/ejercicios/EjerciciosScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 
 @Composable
 fun EjerciciosScreen(
-    onIrADetallesEjercicio : (Int) -> Unit
+    onIrADetallesEjercicio: () -> Unit
 ) {
     Column(modifier = Modifier.fillMaxSize()) {
         Text("Ejercicios")
