+++ app/src/main/java/com/gym/ui/navigation/historial/HistorialRoute.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 data object HistorialRoute
 
 fun NavGraphBuilder.historialDestination(
-    onIrAFormRegistrosPorFecha: (Int) -> Unit
+    onIrAFormRegistrosPorFecha: () -> Unit
 ) {
     composable<HistorialRoute> {
         HistorialScreen(
